agent/pyproject.toml
[project]
name = "opensearch-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.39.4",
    "boto3-stubs==1.39.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain-aws==0.2.28",
    "langchain-community==0.3.27",
    "langgraph==0.5.2",
    "opensearch-py==3.0.0",
    "pydantic>=2.11.7",
    "strands-agents>=1.0.1",
    "tqdm==4.67.1",
    "uvicorn[standard]>=0.35.0",
]
